+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - Logical Interconnect Group
    - Logical Switch Group
    - Network Set
+   - SAN Manager
    - SAS Logical Interconnect Group
    - Scopes
    - Switches