+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Suggested release: v4.4.1 (Unreleased Changes)
+
+#### Bug fixes & Enhancements
+- [#235](https://github.com/HewlettPackard/oneview-sdk-ruby/issues/235) Example file for Synergy LIG fails with 'Interconnect type or Logical Downlink not found!'
+
+
 ## v4.4.0
 
 #### Bug fixes & Enhancements